Assert arguments are valid
Because properties are not private in Python classes, every fool can set every value to whatever property in any class at what ever moment. This seems to be 'pythonic'. Just let it explode if it doesn't fit. Who is going to clean up the mess afterwards is never answered. Python is a horrible language.